+++ src/deviceata.js
@@ -35,7 +35,7 @@ class DeviceAta extends EventEmitter {
         this.restFulConnected = false;
 
         //mqtt
-        const mqtt = account.restFul ?? {};
+        const mqtt = account.mqtt ?? {};
         const mqttEnabled = mqtt.enable || false;
         this.mqttConnected = false;
 
@@ -84,40 +84,9 @@ class DeviceAta extends EventEmitter {
 
                     this.restFul.on('connected', (message) => {
                         this.restFulConnected = true;
-                        this.emit('message', message);
+                        this.emit('success', message);
                     })
-                        .on('debug', (debug) => {
-                            this.emit('debug', debug);
-                        })
-                        .on('error', (error) => {
-                            this.emit('warn', error);
-                        });
-                }
-                const restFul = this.restFulConnected ? this.restFul.update('info', deviceData) : false;
-                const restFul1 = this.restFulConnected ? this.restFul.update('state', deviceState) : false;
-            }
-
-            //MQTT client
-            if (mqttEnabled) {
-                if (!this.mqttConnected) {
-                    this.mqtt = new Mqtt({
-                        host: mqtt.host,
-                        port: mqtt.port || 1883,
-                        clientId: `${mqtt.clientId}_${deviceId}` || `${deviceTypeText}_${deviceName}_${deviceId}`,
-                        prefix: `${mqtt.prefix}/${deviceTypeText}/${deviceName}`,
-                        user: mqtt.user,
-                        passwd: mqtt.pass,
-                        debug: mqtt.debug || false
-                    });
-
-                    this.mqtt.on('connected', (message) => {
-                        this.mqttConnected = true;
-                        this.emit('message', message);
-                    })
-                        .on('subscribed', (message) => {
-                            this.emit('message', message);
-                        })
-                        .on('subscribedMessage', async (key, value) => {
+                        .on('set', async (key, value) => {
                             try {
                                 switch (key) {
                                     case 'Power':
@@ -186,7 +155,7 @@ class DeviceAta extends EventEmitter {
                                         await this.melCloudAta.send(deviceState);
                                         break;
                                     default:
-                                        this.emit('message', `MQTT Received unknown key: ${key}, value: ${value}`);
+                                        this.emit('warn', `MQTT Received key: ${key}, value: ${value}`);
                                         break;
                                 };
                             } catch (error) {
@@ -200,7 +169,114 @@ class DeviceAta extends EventEmitter {
                             this.emit('warn', error);
                         });
                 }
-                const mqtt = this.mqttConnected ? this.mqtt.emit('publish', `Info`, deviceData) : false;
+                const restFul0 = this.restFulConnected ? this.restFul.update('info', deviceData) : false;
+                const restFul1 = this.restFulConnected ? this.restFul.update('state', deviceState) : false;
+            }
+
+            //MQTT client
+            if (mqttEnabled) {
+                if (!this.mqttConnected) {
+                    this.mqtt = new Mqtt({
+                        host: mqtt.host,
+                        port: mqtt.port || 1883,
+                        clientId: `${mqtt.clientId}_${deviceId}` || `${deviceTypeText}_${deviceName}_${deviceId}`,
+                        prefix: `${mqtt.prefix}/${deviceTypeText}/${deviceName}`,
+                        user: mqtt.user,
+                        passwd: mqtt.pass,
+                        debug: mqtt.debug || false
+                    });
+
+                    this.mqtt.on('connected', (message) => {
+                        this.mqttConnected = true;
+                        this.emit('success', message);
+                    })
+                        .on('subscribed', (message) => {
+                            this.emit('message', message);
+                        })
+                        .on('set', async (key, value) => {
+                            try {
+                                switch (key) {
+                                    case 'Power':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Power;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'OperationMode':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.OperationMode;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'SetTemperature':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.SetTemperature;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'DefaultCoolingSetTemperature':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.SetTemperature;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'DefaultHeatingSetTemperature':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.SetTemperature;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'SetFanSpeed':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.SetFanSpeed;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'VaneHorizontal':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.VaneHorizontal;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'VaneVertical':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.VaneVertical;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'HideVaneControls':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Prohibit;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'HideDryModeControl':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Prohibit;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'ProhibitSetTemperature':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Prohibit;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'ProhibitOperationMode':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Prohibit;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    case 'ProhibitPower':
+                                        deviceState[key] = value;
+                                        deviceState.EffectiveFlags = CONSTANTS.AirConditioner.EffectiveFlags.Prohibit;
+                                        await this.melCloudAta.send(deviceState);
+                                        break;
+                                    default:
+                                        this.emit('warn', `MQTT Received key: ${key}, value: ${value}`);
+                                        break;
+                                };
+                            } catch (error) {
+                                this.emit('warn', `MQTT send error: ${error}.`);
+                            };
+                        })
+                        .on('debug', (debug) => {
+                            this.emit('debug', debug);
+                        })
+                        .on('error', (error) => {
+                            this.emit('warn', error);
+                        });
+                }
+                const mqtt0 = this.mqttConnected ? this.mqtt.emit('publish', `Info`, deviceData) : false;
                 const mqtt1 = this.mqttConnected ? this.mqtt.emit('publish', `State`, deviceState) : false;
             }
         })