+++ CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### After update to v2.0.0 RESTFull and MQTT config settings need to be updated
 
+## [2.2.0] - (18.08.2024)
+
+## Changes
+
+- fix MQTT and RESTFul start
+- add control over RESTFul POST JSON Object
+- bump dependencies
+
 ## [2.1.1] - (18.08.2024)
 
 ## Changes